Made it actually work
Converted the values to work with our code instead of the arduino. I also found out that the orientation is World->Sensor and we need Sensor->World, so I changed the rotation function to use the inverse of our quaternion. I also added a statement to detect if the acceleration is pegged and discard the data if it is.